adding React Router, nav bar is set up
